requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
fire==0.4.0
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
mypy==0.931
mypy-extensions==0.4.3
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
termcolor==1.1.0
tomli==2.0.0
typed-ast==1.5.2
typing-extensions==4.0.1
urllib3==1.26.8
zipp==3.7.0
